backend/requirements.txt
openai
openai[datalib]>=1.3.6
tiktoken
uvicorn
aiohttp
fastapi
pydantic
spacy
pymupdf
python-dateutil
python-multipart
anthropic
python-dotenv
beautifulsoup4
gensim
numpy
pandas
selenium
requests
scipy==1.10.1